refactor(analysis-types): extract complexity risk DTO
